refactor(TextInputWrapper): moved to internal
