+++ src/TextInputWithTokens/TextInputWithTokens.tsx
@@ -10,7 +10,7 @@ import {TextInputProps} from '../TextInput'
 import Token from '../Token/Token'
 import {TokenSizeKeys} from '../Token/TokenBase'
 import TextInputInnerVisualSlot from '../_TextInputInnerVisualSlot'
-import TextInputWrapper, {textInputHorizPadding, TextInputSizes} from '../_TextInputWrapper'
+import TextInputWrapper, {textInputHorizPadding, TextInputSizes} from '../internal/components/TextInputWrapper'
 import UnstyledTextInput from '../_UnstyledTextInput'
 
 // eslint-disable-next-line @typescript-eslint/no-explicit-any
